+++ extensions/Splitwise/src/hooks/useCurrentUser.tsx
@@ -1,10 +1,7 @@
 import { useFetch } from "@raycast/utils";
-import { Cache } from "@raycast/api";
 
 import { GetUser } from "../types/get_user.types"; // Types
 import { HEADER } from "./userPreferences";
-
-// const cached = new Cache();
 
 export function GetCurrentUser() {
   const { data, error } = useFetch<GetUser>(`https://secure.splitwise.com/api/v3.0/get_current_user`, {
@@ -14,7 +11,6 @@ export function GetCurrentUser() {
   });
 
   const currentUser = data?.user;
-  // cached.set("currentUser",  JSON.stringify(currentUser));
 
   if (error) {
     console.log(`Error while fetching currrent user: \n ${error}`);