+++ extensions/Splitwise/src/hooks/userPreferences.tsx
@@ -4,7 +4,7 @@ const { personalAccessToken } = getPreferenceValues();
 export const { loadingLimit } = getPreferenceValues();
 
 export const HEADER = {
-    headers: {
-      Authorization: `Bearer ${personalAccessToken}`,
-    },
-  };
+  headers: {
+    Authorization: `Bearer ${personalAccessToken}`,
+  },
+};