+++ extensions/Splitwise/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Update] - 2023-16-09
 - Added new command `List Expenses` which allows to list all expenses and edit them
 - Added validation to `Add Expense` command
+- Changed friend's balance from text to tags
+- Improved API Calls
 
 ## [Update] - 2023-08-09
 - Added support for adding expenses to groups